GLMMS/requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.0.0
Pillow==5.3.0
pytz==2018.7
